Module alesia::napi

Exposes FFI-functions that are exported into dynamic libraries or shared objects.

The functions follow the following general name scheme:

If the function name name_of_the_function is unique across the crate, then alsNameOfTheFunction is the chosen name 1.1 For Clarity’s sake, the function name may be followed by _Struct if function is a struct method.

If the function is a constructor for a struct, i.e, Struct::new then alsNewStruct is the chosen name

Destructors have names of the form alsFreeStruct where Struct::drop is invoked.

Functions exclusive to the Native API begin with alsn and are named in CamelCase.

Wrapper functions will not panic unless underlying engine functions panic, i.e, in case of errors most wrapper functions will simply return (in case of void return type) or return a default value after printing appropriate error messages. Exception. Wrapper functions that require a const c_char* pointer panic if the CStr is not a valid Utf8 String. If the function returns a mutable pointer, then in case of an error a NULL Pointer is returned.

Functions

alsBeginS_Display

Creates the display and begins the game. ResourceSet and World are deallocated when this method returns.

alsBegin_Display

Creates the display and begins the game. ResourceSet, World, and StateListener are deallocated when this method returns.

alsBindInit

alsBindTurn

alsBlank_World

alsCreateStatic

alsDefAnim

FFI for UnitType.def_anim. Tuples have been expanded into individual arguments.

alsFreeResourceSet

Free a leaked reference

alsFreeStateListener

Free a leaked reference

alsFreeUnitType

Free a leaked reference

alsFreeWorld

Free a leaked reference

alsIdList

alsMapFont

alsMapMusic

alsMapSound

alsMapTexture

alsNewResourceSet

alsNewStateListener

alsNewUnitType

alsRegisterUnitType

alsSpawnUnit

alsnGetUnitHealth

Getter for the health of the unit with specified ID. Returns -1.0 on NULL pointer or invalid ID.

alsnGetUnitX

Getter for the x position of the unit with specified ID.

alsnGetUnitY

Getter for the x position of the unit with specified ID.

alsnIsUnitFoe

Returns true if the unit with given id is an enemy unit.

alsnPushAttackOrder

Push an attack Order. See enum Order.

alsnPushMoveOrder

Push a move order. See enum Order.

alsnUnitRef

Get an immutable (i.e, readonly) reference to the unit with the specified id.

alsnVecAt

Return byte value at index elm in Vec<u8>

alsnVecLen

Return length of byte vector.