Function alesia::world::load_world

pub fn load_world(_w: &mut World, fpath: &str) -> bool

Load tile map data from the specified file into the world

_w - The world to load [TileMap] into

fpath - The path to the file containing map data.
Returns true if map data could successfully be loaded, otherwise false.

Binary Format

The file specified by fpath must conform to the following binary format:

First four bytes of the file are exactly [250, 222, 0, 255]
The next byte specifies the width of the map.
The following byte specified the height of the map.
The next w*h bytes, where w and h are map width and height repsectively, comprise map data for each tile.
The next 6 bytes form a mandatory padding (thus must be identically zero).
The remaining section of the file defines game objects, and their position in the world.
Game Objects are encoded as 6 byte sequences that begin with [254, 237]. The third byte defines the game object type.
The fourth byte is the type-parameter for a given game object. The fifth and sixth bytes define the x and y co-ordinates of the game object.

If any of the mandatory components of the format are missing in the file specified, the file is termed as a malformed world file. The term Game Object is merely an abstract construct (with no direct counterpart in the engine) used to allow for a common format of specification for statics, and units. The following table summarizes type and type-parameter relations:

Type (u8)	Parameter

Static (0)	The texture id of the static

Player Unit (1)	The type id of the unit

Enemy Unit (2)	The type id of the unit

Panics

The function panics with appropriate error messages if:

The file could not be found or opened (does not block until file is available)

An I/O Error occurs, and read fails.

Memory allocation of map data fails.